Go through dolphin/os.h

+++ include/dolphin/os/OSAlloc.h
@@ -3,15 +3,15 @@
 
 #include "dolphin/types.h"
 
-typedef int OSHeapHandle;
+typedef s32 OSHeapHandle;
 
 extern volatile OSHeapHandle __OSCurrHeap;
 
-void* OSInitAlloc(void*, void*, int);
+void* OSInitAlloc(void*, void*, s32);
 OSHeapHandle OSCreateHeap(void*, void*);
 OSHeapHandle OSSetCurrentHeap(OSHeapHandle);
 void* OSAllocFromHeap(OSHeapHandle, u32);
-long OSCheckHeap(OSHeapHandle);
+s32 OSCheckHeap(OSHeapHandle);
 void OSFreeToHeap(OSHeapHandle heap, void* ptr);
 
 #define OSAlloc(size) OSAllocFromHeap(__OSCurrHeap, (size))
